+++ Common/Servers/MessageBus.cpp
@@ -29,7 +29,7 @@ namespace
 
     void removeEndpointFromUnorderedVector(std::vector<MessageBusEndpoint *> &vec,const MessageBusEndpoint *to_remove)
     {
-        for(int i=0,total=vec.size(); i<total; ++i)
+        for(int i=0,total=(int)vec.size(); i<total; ++i)
         {
             if(vec[i]==to_remove)
             {
